refactor: bring file names to the same style
