+++ src/index.ts
@@ -1,6 +1,6 @@
 import { startApp } from './app/utils/start-app';
-import { getEnv } from './common/utils/getEnv';
-import { startServer } from './server/utils/startServer';
+import { getEnv } from './common/utils/get-env';
+import { startServer } from './server/utils/start-server';
 
 if (getEnv() === 'prod') {
   startServer();
