+++ src/server/constants.ts
@@ -1,7 +1,7 @@
 import cors from 'cors';
 import { HOST, FRONTEND_PORT } from '../constants';
 import { loadDLL } from './load-DLL';
-import { DLLPathResolve } from './utils/DLLPathResolve';
+import { DLLPathResolve } from './utils/dll-path-resolve';
 
 export const DLL_PATH = DLLPathResolve('libNelderMead.so');
 export const CORS = cors({
