+++ src/server/index.ts
@@ -1,3 +1,3 @@
-import { startServer } from './utils/startServer';
+import { startServer } from './utils/start-server';
 
 startServer();
